python/agents/swe-benchmark-agent/pyproject.toml
[project]
name = "swe-benchmark-agent"
version = "0.1.0"
description = "A software engineering agent for SWE-bench and Terminal-Bench benchmarks using Google ADK"
authors = [
  { name = "Utsav Garg", email = "utsavgarg@google.com" },
  { name = "Dawei Jia", email = "dwkk@google.com" },
]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.10,<3.14"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3 :: Only",
]

dependencies = [
    "swebench @ git+https://github.com/swe-bench/swe-bench.git@v4.1.0#egg=swebench",
    "typer>=0.19.2",
    "datasets>=4.2.0",
    "jinja2>=3.1.5",
    "GitPython>=3.1.45",
    "docker>=7.1.0",
    "google-adk @ file:///usr/local/google/home/txinran/evals/swe/adk-python",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    "google-genai>=1.56.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.23.0",
]

[tool.ruff]
# Point back to the Master template in the root
extend = "../../../pyproject.toml"

[tool.ruff.lint.isort]
# Categorizes the local package for correct import sorting
known-first-party = ["swe_benchmark_agent"]

[tool.ruff.lint]
# Ignores function complication rule
ignore = ["C901", "PLW2901", "PLR0911", "PLR0912", "PLR0915"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]


[tool.hatch.metadata]
# Required to allow the git+https dependency for swebench
allow-direct-references = true

[tool.hatch.build.targets.wheel]
# Explicitly defines the package source for Hatchling
packages = ["swe_benchmark_agent"]

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"